pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kairos-contracts"
version = "0.1.0"
description = "Structured context system for AI agents working in multi-repo ecosystems"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "jsonschema",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.setuptools.packages.find]
include = ["kairos*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
